feat: fetch 10 resutls at a time
