+++ packages/backend/src/routers/projectRouter.ts
@@ -192,12 +192,17 @@ projectRouter.get(
                 typeof req.query.value === 'string'
                     ? req.query.value.toString()
                     : '';
+            const limit: number =
+                typeof req.query.limit === 'string'
+                    ? parseInt(req.query.limit.toString(), 10)
+                    : 10;
             const results: Array<any> =
                 await projectService.searchFieldUniqueValues(
                     req.user!,
                     req.params.projectUuid,
                     req.params.fieldId,
                     value,
+                    limit,
                 );
             res.json({
                 status: 'ok',
